test: 소설 fixture 생성하는 코드 디렉토리 변경
