+++ src/test/java/com/ktb/eatbookappbackend/member/fixture/MemberFixture.java
@@ -14,7 +14,6 @@ public class MemberFixture {
     public static final int TOTAL_ITEMS = 1;
     public static final int TOTAL_PAGES = 1;
     public static final int FAVORITE_COUNT = 5;
-    public static final boolean IS_FAVORITED = true;
 
     public static final int EMPTY_TOTAL_ITEMS = 0;
     public static final int EMPTY_TOTAL_PAGES = 1;
@@ -30,17 +29,7 @@ public class MemberFixture {
         return member;
     }
 
-    public static Novel createNovel() {
-        Novel novel = Novel.builder()
-                .title("Novel Title")
-                .coverImageUrl("coverImageUrl")
-                .summary("Novel Summary")
-                .isCompleted(true)
-                .build();
-        String novelID = UUID.randomUUID().toString();
-        ReflectionTestUtils.setField(novel, "id", novelID);
-        return novel;
-    }
+
 
     public static Bookmark createBookmark(Novel novel, Member member) {
         return Bookmark.builder()
